requirements.txt
streamlit
pandas
python-docx
PyPDF2
reportlab
openai-whisper
streamlit-webrtc
av
